backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
Pillow==10.4.0
qrcode[pil]==7.4.2

